requirements.txt
pytorch-lightning>=1.4.4
tqdm~=4.64.1
bcolz-zipline
prettytable~=3.5.0
menpo~=0.11.0
mxnet~=1.2.0
opencv-python~=4.6.0.66
torch~=1.13.0
numpy~=1.23.4
pandas~=1.5.1
torchvision~=0.9.0
pillow~=9.2.0
scipy~=1.9.3
matplotlib~=3.5.3
scikit-learn~=1.1.3
scikit-image~=0.19.2